requirements.txt
Flask==2.3.2
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.25.2
pandas==2.1.1
opencv-python==4.8.1.78
Pillow==10.0.1
ultralytics==8.0.188
transformers==4.34.1
torch==2.0.1
requests==2.31.0
 